requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==4.3.0
pkg-resources==0.0.0
plaid-python==2.3.4
pluggy==0.8.0
py==1.7.0
pycrypto==2.6.1
pytest==4.0.1
pytest-cov==2.6.0
requests==2.20.1
rncryptor==3.2.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
